Перенос функции лайков в отдельный файл
